refactor(cmd): split unified authbridge binary into 3 mode-specific binaries
The unified `cmd/authbridge/` binary picked its mode at runtime via a
`--mode` flag, but operators always pin the mode via the per-pod
config the operator generates — there's no real runtime selection
happening. The unified shape also caused the recent regression where
`Dockerfile.proxy` silently built the wrong (lite) binary for the
proxy-sidecar combined image, dropping the parser plugins and breaking
abctl observability for proxy-sidecar deployments.

This PR replaces the runtime `--mode` selection with build-time
selection: three mode-specific binaries, one per deployment shape.
Each binary's `main.go` is small and unambiguous (~270 LOC), no
per-mode if/else, and imports exactly the listener and plugins it
needs.

New layout under `cmd/`:

  authbridge-proxy/   proxy-sidecar mode (default). HTTP forward +
                     reverse proxies. Full plugin set including the
                     a2a/mcp/inference parsers, so abctl session
                     events render protocol context.
  authbridge-envoy/   envoy-sidecar mode. ext_proc gRPC server hooked
                     into Envoy. Full plugin set.
  authbridge-lite/    proxy-sidecar mode, lite plugin set (auth gates
                     only — jwt-validation + token-exchange — parsers
                     dropped) for size-optimized deployments.

Each binary lives in its own directory with main.go + go.mod + Dockerfile
+ entrypoint.sh. The CI matrix in .github/workflows/build.yaml now
publishes three images: `authbridge` (proxy combined, default),
`authbridge-envoy` (envoy combined), `authbridge-lite` (proxy combined,
lite). `proxy-init` continues to ship unchanged.

Listener implementations move from cmd/authbridge/listener/ into authlib/
to match the existing forwardproxy/reverseproxy layout — the binaries
no longer own listener code, only main.go composition. authlib/listener/
now has all four listeners (extauthz, extproc, forwardproxy, reverseproxy)
and the binaries import only what they need.

The waypoint mode constant is preserved in authlib/config and the
extauthz listener is preserved in authlib/listener/extauthz, but no
binary registers it currently — waypoint deployments need to wait for
a follow-up PR to bring it back as its own binary if there's demand.

Side effects:
* go.mod files bump to Go 1.25 (envoyproxy/go-control-plane requires it
  after authlib absorbed the listener code).
* Dockerfile golang base bumps from 1.24-alpine to 1.25-alpine to match.
* CLAUDE.md (root and authbridge/) refreshed to describe the new
  three-binary layout, the new image set, and the new code conventions.
* Operator (kagenti-operator) needs no changes — it references
  `authbridge` and `authbridge-envoy` images by name; their content
  changes (now built from the new mode-specific dirs) but the image
  names + ports + entrypoint contracts are stable.
* `authbridge-lite` is a new image not referenced by any current
  consumer; it's available for future size-constrained deployments.

Verified locally: all three images build cleanly, all three load into
kind, weather-agent in proxy-sidecar mode boots with all 5 plugins
registered, abctl receives session events on user requests.
