+++ authbridge/cmd/authbridge-envoy/entrypoint.sh
@@ -2,12 +2,12 @@
 set -eu
 
 # AuthBridge envoy-sidecar combined entrypoint with process supervision.
-# Manages: spiffe-helper (optional), authbridge (ext_proc), envoy.
+# Manages: spiffe-helper (optional), authbridge-envoy (ext_proc), envoy.
 #
 # Startup order:
 #   1. spiffe-helper (background, only when SPIRE_ENABLED=true)
-#   2. authbridge (background) — gRPC ext_proc listener
-#   3. envoy (background) — calls authbridge over ext_proc
+#   2. authbridge-envoy (background) — gRPC ext_proc listener
+#   3. envoy (background) — calls authbridge-envoy over ext_proc
 #
 # Process management: PID 1 (this shell) supervises every long-running
 # critical process. If any critical process exits, the others are killed
@@ -32,12 +32,12 @@ if [ "${SPIRE_ENABLED:-}" = "true" ]; then
   CRITICAL_PIDS="$CRITICAL_PIDS $!"
 fi
 
-# --- Phase 2: authbridge (ext_proc gRPC server) ---
-echo "[entrypoint] Starting authbridge..."
-/usr/local/bin/authbridge "$@" &
+# --- Phase 2: authbridge-envoy (ext_proc gRPC server) ---
+echo "[entrypoint] Starting authbridge-envoy..."
+/usr/local/bin/authbridge-envoy "$@" &
 CRITICAL_PIDS="$CRITICAL_PIDS $!"
 
-# Give authbridge a moment to bind the gRPC listener before Envoy connects
+# Give authbridge-envoy a moment to bind the gRPC listener before Envoy connects
 sleep 2
 
 # --- Phase 3: Envoy ---
