+++ authbridge/cmd/authbridge-proxy/entrypoint.sh
