--- authbridge/cmd/authbridge/Dockerfile.envoy
@@ -1,47 +1,0 @@
-# AuthBridge envoy-sidecar combined image — Envoy + authbridge (ext_proc)
-# + spiffe-helper in a single container.
-#
-# Spiffe-helper starts conditionally based on the SPIRE_ENABLED env var
-# (set by the operator when the workload opts into SPIRE identity).
-#
-# Image lineage: replaces the older `authbridge` (envoy + everything,
-# UBI9-micro base ~140 MB combined) and `authbridge-unified` images.
-#
-# Build context: ./authbridge (needs access to authlib/ and cmd/authbridge/)
-
-# Stage 1: Build authbridge binary
-FROM golang:1.24-alpine@sha256:8bee1901f1e530bfb4a7850aa7a479d17ae3a18beb6e09064ed54cfd245b7191 AS builder
-
-RUN apk add --no-cache git
-
-WORKDIR /app
-
-COPY authlib/ authlib/
-COPY cmd/authbridge/ cmd/authbridge/
-
-ENV GOWORK=off
-RUN cd cmd/authbridge && CGO_ENABLED=0 GOOS=linux go build -o /authbridge .
-
-# Stage 2: Get Envoy binary
-FROM docker.io/envoyproxy/envoy:v1.37.1 AS envoy-source
-
-# Stage 3: Get spiffe-helper binary
-FROM ghcr.io/spiffe/spiffe-helper:0.11.0 AS spiffe-source
-
-# Stage 4: Runtime — UBI9-micro (glibc-compatible, ~24 MB, has bash)
-FROM registry.access.redhat.com/ubi9/ubi-micro@sha256:2173487b3b72b1a7b11edc908e9bbf1726f9df46a4f78fd6d19a2bab0a701f38
-
-COPY --from=builder /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
-COPY --from=envoy-source --chmod=755 /usr/local/bin/envoy /usr/local/bin/envoy
-COPY --from=spiffe-source --chmod=755 /spiffe-helper /usr/local/bin/spiffe-helper
-COPY --from=builder --chmod=755 /authbridge /usr/local/bin/authbridge
-COPY --chmod=755 cmd/authbridge/entrypoint-envoy.sh /usr/local/bin/entrypoint.sh
-
-# Envoy writes hot-restart lock files and admin socket here
-RUN mkdir -p /tmp/envoy && chmod 775 /tmp/envoy
-
-USER 1337
-
-EXPOSE 15123 15124 9901 9090
-
-ENTRYPOINT ["/usr/local/bin/entrypoint.sh"]
